Sync. with bz-puppeteer

+++ wrap.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
 Xvfb :99 -screen 0 1280x1024x16 & node index --docker $@
-cp /root/Downloads/* /var/boozang/
+if [ -e /root/Downloads ]
+  cp /root/Downloads/* /var/boozang/
+fi
